Function to check user-specified grid for FFTKDE
Adds a Cython function which verifies that a grid is sorted. Verifies that a grid is sorted when used with FFTKDE. No option to turn this check off, since we assume the number of grid points is much smaller than the number of data points, and the bottleneck is iterating over the data points.     * Adds unit tests and regression tests.